+++ src/components/artist-actions.ts
@@ -316,8 +316,8 @@ class ArtistActions extends FavActionsBase {
 
       /* style ha-icon-button controls in tracks grid: icon size, title text */
       .albums-grid > ha-icon-button[slot="icon-button"] {
-        --mdc-icon-button-size: 24px;
-        --mdc-icon-size: 20px;
+        --ha-icon-button-size: 24px;
+        --ha-icon-size: 20px;
         vertical-align: top;
         padding: 0px;
       }