+++ src/components/audiobook-actions.ts
@@ -255,8 +255,8 @@ class AudiobookActions extends FavActionsBase {
 
       /* style ha-icon-button controls in tracks grid: icon size, title text */
       .chapters-grid > ha-icon-button[slot="icon-button"] {
-        --mdc-icon-button-size: 24px;
-        --mdc-icon-size: 20px;
+        --ha-icon-button-size: 24px;
+        --ha-icon-size: 20px;
         vertical-align: top;
         padding: 0px;
       }